pyproject.toml
[project]
name = "eink-info-bw"
version = "0.1.0"
description = "Web-managed eInk photo frame for Raspberry Pi Zero 2 W"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "jinja2>=3.1.6",
  "pillow>=11.3.0",
  "python-multipart>=0.0.20",
  "uvicorn[standard]>=0.35.0",
]

[tool.uv]
package = false
